requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.2
django
djangorestframework
celery
redis
requests